+++ src/enums/browser-theme.ts
@@ -1,4 +1,4 @@
-export enum Theme {
+export enum BrowserTheme {
     Light = 'light',
     Dark = 'dark',
 }